Reuse the context and state in js.VU as much as possible
This took way longer then I anticipated mostly because of ... how
strange some things work.

The basic idea came from that for each iteration we constantly remake the
iterations which looked (and probably is) the lowest hanging fruit in k6
about creating too much objects.

Unfortunately because of how .. tags work and that iteration is a tag I
need to still copy the tags for GetTrail .. (it gets copied either way
for other metrics).

When running with a empty body did this change the amount of cpu taken
by scanobject and decreased it from 7.98s to 5.81s (while most other not
GC related had difference withing 0.3s).

This is very small step ... having tags being immutable so we don't have
to constantly copy them will likely be much more significant or if we
can reduce the calls to runtime.nanotime which are literally now the top
CPU usage with empty iterations and are the top CPU usage for simple
scripts even before that :D.