Fix externally controlled executor not pausing
The problem was that pausing used gracefulStop which in turns wait for
the VU used by handle to be returned. The fix is to have another context
that is used for the actual runContext of the VU so we can cancel it
when we get gracefully stopped.

I tried to simplify (and speed up) the looping by only going through all
the channels and context stuff if there was a change, otherwise we use a
int32 and some atomic.LoadInt32 to see that "there is no change" and we
directly start a new iteration.

Additionally now the getVU error is propagetted by the vu handle's
start, which is probably badly handled in this commit.

The speed up is insignficant, by my proffiling we were sepending around
150-250ms (of 30s) in the mutex locking/channel reading now we spend 10ms(of
30s) in loading that int32.